#104 Перенес чарт в отдельный компонент
