+++ frontend/src/app/profile/components/profile-projects/profile-projects-chart/profile-projects-chart.component.scss
@@ -1,0 +1,3 @@
+:host {
+  display: block;
+}
