+++ frontend/src/app/profile/components/profile-projects/profile-projects.component.html
@@ -1,0 +1,1 @@
+<app-profile-projects-chart [projects]="projects"></app-profile-projects-chart>
