+++ frontend/src/app/profile/profile.module.ts
@@ -4,6 +4,7 @@ import { SharedModule } from '../shared/shared.module';
 import { ProfileFormComponent } from './components/profile-form/profile-form.component';
 import { ProfileHistoryComponent } from './components/profile-history/profile-history.component';
 import { ProfilePageComponent } from './components/profile-page.component';
+import { ProfileProjectsChartComponent } from './components/profile-projects/profile-projects-chart/profile-projects-chart.component';
 import { ProfileProjectsComponent } from './components/profile-projects/profile-projects.component';
 import { ProfileSubscriptionsComponent } from './components/profile-subscriptions/profile-subscriptions.component';
 
@@ -12,7 +13,8 @@ const components = [
   ProfileHistoryComponent,
   ProfileSubscriptionsComponent,
   ProfileFormComponent,
-  ProfileProjectsComponent
+  ProfileProjectsComponent,
+  ProfileProjectsChartComponent
 ];
 
 @NgModule({
